+++ docs/changelog.rst
@@ -5,6 +5,17 @@ Changelog
 master
 ======
 
+Other Changes
+-------------
+
+* The ``canonical_url`` option was deprecated in favor of Sphinx's ``html_baseurl``.
+
+New Features
+------------
+
+* New theme option to enable anonymous ip addresses when using Google Analytics (#889)
+
+
 v0.5.0
 ======